task5

+++ resources/item.py
@@ -6,6 +6,7 @@ from models import ItemModel
 import uuid
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from db import items
 from schemas import ItemSchema
 from schemas import ItemUpdateSchema
@@ -14,18 +15,29 @@ blp = Blueprint("Items", __name__, description="Operations on items")
 
 @blp.route("/item/<string:item_id>")
 class Item(MethodView):
-    @blp.response(200, ItemSchema)
+    @jwt_required()
     def get(self, item_id):
+        item = ItemModel.query.get_or_404(item_id)
+        return item
+
+    @jwt_required()
+    def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+        
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
         return {"message": "Item deleted."}
-
-    def delete(self, item_id):
-        item = ItemModel.query.get_or_404(item_id)
-        raise NotImplementedError("Deleting an item is not implemented.")
     @blp.route("/item")
     class ItemList(MethodView):
+        @jwt_required()
+        @blp.response(200, ItemSchema(many=True))
+        def get(self):
+            return ItemModel.query.all()
+
+        @jwt_required()
         @blp.arguments(ItemUpdateSchema)
         @blp.response(200, ItemSchema)
         def put(self, item_data, item_id):
@@ -41,6 +53,7 @@ class Item(MethodView):
 
             return item
 
+        @jwt_required(fresh=True)
         @blp.arguments(ItemSchema)
         @blp.response(201, ItemSchema)
         def post(self, item_data):
